College Mailroom Database
Created an organized database that efficiently distribute mail
throughout college campuses
Used SQL and PHP for back-end operation. Use HTML to create a user
interface for analysis of the database.
The organization is for logging mail, notifying the student of a
package and logs a package out once after pick up.